classly/backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
supabase>=2.27.0
websockets>=12.0
requests==2.31.0
selenium==4.15.0
webdriver-manager==4.0.1
beautifulsoup4==4.12.2
gunicorn==21.2.0
# AI Agent dependencies
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.20
openai>=1.0.0
PyJWT>=2.8.0
# Google Calendar integration
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.1.0
pytz>=2023.3
